# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Airline Booking System API
  version: 1.1.0
  description: |
    Contract for the Airline Booking System gateway (http://localhost:8080). All calls go through
    the API Gateway; it validates JWTs, enforces ADMIN roles and routes to the micro-services.
    Protected endpoints require `Authorization: Bearer <jwt>`. The gateway injects the
    `X-User-Email` and `X-User-Role` headers for the services, so clients do not send them.
    All errors use a uniform body: {"timestamp","status","error","message","path"}.
servers:
  - url: http://localhost:8080
    description: API Gateway (the only backend URL a frontend should call)
security:
  - bearerAuth: []
...
paths:
...
  /api/flights/search:
    get:
      tags: [Flights]
      security: []
      summary: Search active flights
      parameters:
        - { name: origin, in: query, required: true, schema: { type: string, minLength: 3, maxLength: 3 }, description: Departure airport code }
        - { name: destination, in: query, required: true, schema: { type: string, minLength: 3, maxLength: 3 }, description: Arrival airport code }
        - { name: date, in: query, required: true, schema: { type: string, format: date }, description: Departure date }
        - { name: passengers, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 }, description: Required available seats }
        - { name: airline, in: query, required: false, schema: { type: string }, description: Optional airline filter }
        - { name: maxPrice, in: query, required: false, schema: { type: number, format: double }, description: Optional max fare filter }
        - { name: maxDurationMinutes, in: query, required: false, schema: { type: integer }, description: Optional max flight duration filter }
      responses:
        '200':
          description: Matching flights
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Flight' }
  /api/flights/{id}:
    get:
      tags: [Flights]
      security: []
      summary: Retrieve flight details by ID
      parameters:
        - { $ref: '#/components/parameters/FlightId' }
      responses:
        '200':
          description: Flight object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Flight' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/flights/admin:
    post:
      tags: [Flights]
      summary: Create a flight (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FlightRequest' }
      responses:
        '201':
          description: Flight created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Flight' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/flights/admin/{id}:
    put:
      tags: [Flights]
      summary: Update a flight (ADMIN)
      parameters:
        - { $ref: '#/components/parameters/FlightId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FlightRequest' }
      responses:
        '200':
          description: Flight updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Flight' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/flights/admin/{id}/delay:
    post:
      tags: [Flights]
      summary: Delay a flight by N minutes (ADMIN)
      parameters:
        - { $ref: '#/components/parameters/FlightId' }
        - { name: minutes, in: query, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Delayed flight
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Flight' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/flights/admin/{id}/cancel:
    post:
      tags: [Flights]
      summary: Cancel a flight (deactivates it) (ADMIN)
      parameters:
        - { $ref: '#/components/parameters/FlightId' }
      responses:
        '200':
          description: Cancelled flight
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Flight' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  # Airport catalogue (ADMIN)
...
  /api/passengers:
    post:
      tags: [Passengers]
      summary: Create a traveller profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PassengerRequest' }
      responses:
        '201':
          description: Traveller created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Passenger' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/passengers/{id}:
    put:
      tags: [Passengers]
      summary: Update a traveller profile (owner only)
      parameters:
        - { $ref: '#/components/parameters/PassengerId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PassengerRequest' }
      responses:
        '200':
          description: Traveller updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Passenger' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/passengers/{id}/points:
    post:
      tags: [Passengers]
      summary: Add frequent-flyer points (owner only)
      parameters:
        - { $ref: '#/components/parameters/PassengerId' }
        - { name: points, in: query, required: true, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Updated traveller with new frequentFlyerPoints
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Passenger' }
        '404': { $ref: '#/components/responses/NotFound' }
  # ----------------------------- Booking Service -----------------------------
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    PassengerId: { name: id, in: path, required: true, schema: { type: integer, format: int64 }, description: Passenger identifier }
    FlightId: { name: id, in: path, required: true, schema: { type: integer, format: int64 }, description: Flight identifier }
...
  schemas:
    Error:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }
...
    Flight:
      type: object
      properties:
        id: { type: integer, format: int64 }
        flightNumber: { type: string }
        origin: { type: string, minLength: 3, maxLength: 3 }
        destination: { type: string, minLength: 3, maxLength: 3 }
        departureTime: { type: string, format: date-time }
        arrivalTime: { type: string, format: date-time }
        fare: { type: number, format: double }
        availableSeats: { type: integer }
        totalSeats: { type: integer }
        airline: { type: string }
        aircraftCode: { type: string }
        status: { type: string, enum: [SCHEDULED, DELAYED, CANCELLED] }
        delayMinutes: { type: integer }
        active: { type: boolean }
    FlightRequest:
      type: object
      required: [flightNumber, origin, destination, departureTime, arrivalTime, fare, availableSeats, totalSeats, airline]
      properties:
        flightNumber: { type: string }
        origin: { type: string, minLength: 3, maxLength: 3 }
        destination: { type: string, minLength: 3, maxLength: 3 }
        departureTime: { type: string, format: date-time }
        arrivalTime: { type: string, format: date-time }
        fare: { type: number, format: double, minimum: 0 }
        availableSeats: { type: integer, minimum: 0 }
        totalSeats: { type: integer, minimum: 1 }
        airline: { type: string }
        aircraftCode: { type: string }
...
    Passenger:
      type: object
      properties:
        id: { type: integer, format: int64 }
        ownerEmail: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string, format: date }
        phone: { type: string }
        documentNumber: { type: string }
        passportNationality: { type: string }
        passportExpiryDate: { type: string, format: date }
        frequentFlyerPoints: { type: integer }
        savedTraveler: { type: boolean }
    PassengerRequest:
      type: object
      required: [firstName, lastName, dateOfBirth]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        dateOfBirth: { type: string, format: date }
        phone: { type: string }
        documentNumber: { type: string }
        passportNationality: { type: string }
        passportExpiryDate: { type: string, format: date }
        savedTraveler: { type: boolean, default: true }
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: The authenticated user lacks the ADMIN role for this route
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...